requirements.txt
pyside6
pyqtdarktheme
numpy
pyopengl
pillow
pyusb
libusb-package
appdirs
pyinstaller
keyboard
pyserial
